#3 [refactor] 기능별 폴더 안 도메인별 package 제거
