+++ god-care-api/src/main/java/com/godcare/api/exception/CategoryNotFoundException.java
@@ -1,0 +1,4 @@
+package com.godcare.api.exception;
+
+public class CategoryNotFoundException extends RuntimeException{
+}
